Move CreateAuthPreference implementation to auth.Server
The original implementation of `(*clusterconfigv1.Service).CreateAuthPreference`
was meant to be called during the initialization of new clusters
to validate the configured auth preference. However, during init
`auth.Server.CreateAuthPreference` is called, which invoked
`(*local.ClusterConfigurationService) CreateAuthPreference` via the
embedded Services interface instead.

To remedy this `(*clusterconfigv1.Service).CreateAuthPreference`
was moved to `(*auth.Server).CreateAuthPreference`. The authorization
checks were also removed since this is an internal only API that
cannot be called outside of the auth service.
