+++ lib/auth/clusterconfig/clusterconfigv1/service_test.go
@@ -47,114 +47,6 @@ import (
 	"github.com/gravitational/teleport/lib/tlsca"
 )
 
-func TestCreateAuthPreference(t *testing.T) {
-	authRoleContext, err := authz.ContextForBuiltinRole(authz.BuiltinRole{
-		Role:     types.RoleAuth,
-		Username: string(types.RoleAuth),
-	}, nil)
-	require.NoError(t, err, "creating auth role context")
-
-	cases := []struct {
-		name       string
-		modules    modules.Modules
-		authorizer authz.Authorizer
-		preference func(p types.AuthPreference)
-		assertion  func(t *testing.T, created types.AuthPreference, err error)
-	}{
-		{
-			name: "unauthorized built in role",
-			authorizer: authz.AuthorizerFunc(func(ctx context.Context) (*authz.Context, error) {
-				return authz.ContextForBuiltinRole(authz.BuiltinRole{
-					Role:     types.RoleProxy,
-					Username: string(types.RoleProxy),
-				}, nil)
-			}),
-			assertion: func(t *testing.T, created types.AuthPreference, err error) {
-				assert.Nil(t, created)
-				require.True(t, trace.IsAccessDenied(err), "got (%v), expected proxy role to be prevented from creating auth preferences", err)
-			},
-		},
-		{
-			name: "authorized built in auth",
-			authorizer: authz.AuthorizerFunc(func(ctx context.Context) (*authz.Context, error) {
-				return authRoleContext, nil
-			}),
-			assertion: func(t *testing.T, created types.AuthPreference, err error) {
-				require.NoError(t, err, "got (%v), expected auth role to create auth mutator", err)
-				require.NotNil(t, created)
-			},
-		},
-		{
-			name: "creation prevented when hardware key policy is set in open source",
-			authorizer: authz.AuthorizerFunc(func(ctx context.Context) (*authz.Context, error) {
-				return authRoleContext, nil
-			}),
-			preference: func(p types.AuthPreference) {
-				pp := p.(*types.AuthPreferenceV2)
-				pp.Spec.RequireMFAType = types.RequireMFAType_HARDWARE_KEY_PIN
-			},
-			assertion: func(t *testing.T, created types.AuthPreference, err error) {
-				assert.Nil(t, created)
-				require.True(t, trace.IsAccessDenied(err), "got (%v), expected hardware key policy to be rejected in OSS", err)
-			},
-		},
-		{
-			name: "creation allowed when hardware key policy is set in enterprise",
-			authorizer: authz.AuthorizerFunc(func(ctx context.Context) (*authz.Context, error) {
-				return authRoleContext, nil
-			}),
-			modules: &modules.TestModules{TestBuildType: modules.BuildEnterprise},
-			preference: func(p types.AuthPreference) {
-				pp := p.(*types.AuthPreferenceV2)
-				pp.Spec.RequireMFAType = types.RequireMFAType_HARDWARE_KEY_PIN
-			},
-			assertion: func(t *testing.T, created types.AuthPreference, err error) {
-				require.NoError(t, err, "got (%v), expected auth role to create auth mutator", err)
-				require.NotNil(t, created)
-			},
-		},
-		{
-			name: "creation prevented when hardware key policy is set in open source",
-			authorizer: authz.AuthorizerFunc(func(ctx context.Context) (*authz.Context, error) {
-				return authRoleContext, nil
-			}),
-			preference: func(p types.AuthPreference) {
-				p.SetDeviceTrust(&types.DeviceTrust{
-					Mode: constants.DeviceTrustModeRequired,
-				})
-			},
-			assertion: func(t *testing.T, created types.AuthPreference, err error) {
-				assert.Nil(t, created)
-				require.True(t, trace.IsBadParameter(err), "got (%v), expected device trust mode conflict to prevent creation", err)
-			},
-		},
-	}
-
-	for _, test := range cases {
-		t.Run(test.name, func(t *testing.T) {
-			if test.modules != nil {
-				modules.SetTestModules(t, test.modules)
-			}
-
-			var opts []serviceOpt
-			if test.authorizer != nil {
-				opts = append(opts, withAuthorizer(test.authorizer))
-			}
-
-			env, err := newTestEnv(opts...)
-			require.NoError(t, err, "creating test service")
-
-			pref := types.DefaultAuthPreference()
-			if test.preference != nil {
-				test.preference(pref)
-			}
-
-			created, err := env.CreateAuthPreference(context.Background(), pref)
-			test.assertion(t, created, err)
-		})
-	}
-}
-
 func TestGetAuthPreference(t *testing.T) {
 	cases := []struct {
 		name       string
@@ -1437,9 +1329,6 @@ func (failingConfigService) GetSessionRecordingConfig(ctx context.Context) (type
 	return types.DefaultSessionRecordingConfig(), nil
 }
 
-func (failingConfigService) CreateAuthPreference(ctx context.Context, preference types.AuthPreference) (types.AuthPreference, error) {
-	return nil, errors.New("fail")
-}
 func (failingConfigService) UpdateAuthPreference(ctx context.Context, preference types.AuthPreference) (types.AuthPreference, error) {
 	return nil, errors.New("fail")
 }
@@ -1888,7 +1777,7 @@ func newTestEnv(opts ...serviceOpt) (*env, error) {
 	ctx := context.Background()
 	var defaultPreference types.AuthPreference
 	if cfg.defaultAuthPreference != nil {
-		defaultPreference, err = cfg.service.CreateAuthPreference(ctx, cfg.defaultAuthPreference)
+		defaultPreference, err = cfg.service.UpsertAuthPreference(ctx, cfg.defaultAuthPreference)
 		if err != nil {
 			return nil, trace.Wrap(err, "creating default auth mutator")
 		}
